Create separate class to group common objects related to buildRecovery() method
"StorageEngine", "TransactionIdStore" and "LogFiles" were moved to the new class (CoreBuildStore) as they are common in three methods (buildRecovery(), buildKernel() and buildTransactionLogs()). This specific commit addresses the long parameter list in buildRecovery() and CoreBuildStore will be used to address long parameter lists in the other two methods in subsequent commits. All tests pass.